requirements.txt
robotframework==6.1.1
robotframework-tidy==4.8.0
robotframework-soaplibrary==1.3
python-dotenv==1.0.0
